requirements.txt
streamlit==1.24.0
openai==1.10.0
python-dotenv==1.0.0
streamlit-authenticator==0.2.1
PyYAML==6.0
python-docx
